+++ packages/chatbot-server-mongodb-public/src/processors/makeStepBackUserQuery.eval.ts
@@ -4,7 +4,8 @@ import {
   makeStepBackUserQuery,
   StepBackUserQueryMongoDbFunction,
 } from "./makeStepBackUserQuery";
-import { Message, ObjectId, updateFrontMatter } from "mongodb-chatbot-server";
+import { Message, updateFrontMatter } from "mongodb-chatbot-server";
+import { ObjectId } from "mongodb-rag-core/mongodb";
 import { MongoDbTag } from "../mongoDbMetadata";
 import {
   OPENAI_PREPROCESSOR_CHAT_COMPLETION_DEPLOYMENT,
