+++ packages/chatbot-server-mongodb-public/tsconfig.build.json
@@ -1,4 +1,4 @@
 {
   "extends": "./tsconfig.json",
-  "exclude": ["**/*.test.ts", "**/*.eval.ts"]
+  "exclude": ["**/*.test.ts"]
 }
